+++ maintainer/make-release.sh
@@ -50,7 +50,7 @@ if ! cargo msrv --help >/dev/null; then
     exit 1
 fi
 
-if ! cargo msrv --min 1.64.0 --max 1.64.0 >/dev/null; then
+if ! cargo msrv --min 1.70.0 --max 1.70.0 >/dev/null; then
     echo "Minimum supported Rust version doesn't match avertised one."
     exit 1
 fi