+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - when calling `BinRemapper::new`, the limits are now checked for overlaps, in
   which case a new error is returned
+- changed the type `ParseBinRemapperError` to allow capturing errors from
+  `BinRemapper::new`
 
 ## [0.6.3] - 12/12/2023