+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - changed the type `ParseBinRemapperError` to allow capturing errors from
   `BinRemapper::new`
 
+### Removed
+
+- removed the type `SubGrid` that was exported in the CAPI as the type
+  `pineappl_subgrid`. This type was not used anymore and was a left-over of the
+  changes in the previous version.
+
 ## [0.6.3] - 12/12/2023
 
 ### Added